requirements.txt
dataprep=0.4.5
numpy==1.23.3
pandas==1.5.0
scikit-learn=1.1.2
matplotlib==3.6.0
dataprep==0.4.5
typing=3.7.4.3
hydra-core==1.2.0
DataSynthesizer==0.1.11